pyproject.toml
[tool.poetry]
name = "online-cinema"
version = "0.1.0"
description = "FastAPI backend for an online cinema platform"
authors = ["Misha"]
readme = "README.md"
packages = [{ include = "online_cinema", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
aiosqlite = "^0.20.0"
alembic = "^1.14.0"
pydantic-settings = "^2.6.0"
email-validator = "^2.2.0"
python-multipart = "^0.0.12"
pyjwt = "^2.10.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
redis = "^5.2.0"
celery = "^5.4.0"
httpx = "^0.28.0"
minio = "^7.2.12"
boto3 = "^1.35.0"
stripe = "^11.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
respx = "^0.22.0"
types-passlib = "^1.7.7.20240819"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov=online_cinema --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ASYNC", "RUF"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
plugins = []
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
strict_equality = true
exclude = ["tests"]
